Implemented the main structure of a state machine control logic for Stage Pi.
The code requires testing, and main.py and config.py need to be refactored to run a set-up
function one time before the logic runs, so that the state machine object can be
instantiated.

coretools.py is getting quite long, so some refactoring into separate files may be needed
later on.