Fix bug in stagepilogic_tests: fault simulation while loop did not reset faults to zero
The fault iteration while loop was taking the water model through every fault state, but then
leaving it in the last fault state.

This would have been fine, except that the while loop is contained within an outer for loop
which tries each of several water level scenarios. Only the first scenario in the for loop
was being properly tested. This became obvious when fault simulation was disabled and only
one "Testing the "no-faults" state." message was printed where three or four should have been.

To fix this, the water model's resetFaults() method is now called before each run of the
fault iteration while loop. This makes sure it is back at fault state 0 (no faults) on each
iteration of the outer for loop.