Gracefully handle inability to fetch readings in StagePiReadingsParser
StagePiReadingsParser should not raise an exception at initailisation in
the case where it cannot fetch some or all readings. (E.g.: database is
offline)

Instead, each method of the parser should raise a ValueError if the parser
does not have enough readings data to return a valid result.

In this initial fix, the parser insists on more readings than it strictly
requires, but it should cope well with the case where only locally-
originated readings are available. (If logiccoretools is changed to
provide local readings.)