Fixed a design bug whereby Stage Pi Logic would only attempt to control devices when entering a new control logic state, and not continually.
This bug had the effect that the Stage Pi logic needed to start after the NAS and any
devices the logic wanted to control, and was not robust to temporary failures of the
NAS or the devices it wanted to control.

This commit changes the design so that the logic attempts to control devices on each
execution of doLogic, rather than only when entering the state.

Device control is still also attempted upon entering a new state, so that the control
will be timely and not delayed until the next execution of doLogic (which could be 30
seconds away, or more). As such, doLogic does not attempt to control devices if it is
going to put the logic into a new state, because that would result in pointlessly
issuing two control attempts in quick succession, the first of which (from the current
state) would be immediately superseded by the second (from the new state).