Change Temporary Top Up failsafe_end_time to allow 1 hour for top-up.
The original value of 15:30 assumed only the G1 butts group would be filling and that
half an hour would surely be long enough to fill the butts. However, in reality, it
is likely that the G2 and G3 butts groups will also fill, due to the gate valves
having been manually opened, and the manual valve to the rest of the system having
been manually closed. Changing the value to 16:00 will provide an hour, which ought to
be enough.

(N.B. The failsafe_end_time defines a last-resort cut-off time. Normally the top-up
should terminate before this time, due to the water level reaching stop_level.)