Fix bugs in temptopup_control_logic function and corresponding setup function
The temptopup_control_logic_setup function was removed, because the setup function does
not receive the devices dictionary, which means it can't initialise the solenoid device
reference, which meant it had to start the logic without the solenoid, causing a
spurious error message.

Instead, the CSM is initialised by the main control logic function in a conditional
statement.

Other bugs fixed in this commit include a failure to set software_status in all
conditional branches, and a failure to avoid running the CSM's doLogic if the solenoid
object reference had not been initialised.