Temp Top Up logic should fall back to Idle state if readings can't be parsed
The Temp Top Up logic was programmed to 'stall' in its current state if
readings couldn't be parsed; a behaviour copied from Stage Pi Logic. But
it is more appropriate to fall back to Idle state in the case of the Temp
Top up logic, so this commit changes the behaviour to do that instead.