Proper reading interval handling for GenericControlState and subclasses.
The doLogic() method of a control state (subclass of ControlStateABC) should return
the preferred reading interval of that state if no state transition occurs, or that of
the new state if a transition occurs.

But, in error, GenericControlState's implementation of doLogic omitted to return
anything other than the present state's preferred reading interval.

To fix this:

ControlStateMachineABC gains a new getPreferredReadingInterval() method, which returns
the preferred reading interval of the state machine's current state.

Then, GenericControlState's doLogic() can use the CSM's getPreferredReadingInterval()
to get the correct reading interval after running its subclass' stateTransition().

The subclass' stateTransition() method is therefore relieved of any responsibility for
determining or returning an appropriate reading interval value.