Fix temp-top-up override parsing fault, standardise device control strings, docs update
Fault fixed:
logiccoretools.get_state() returns tuple, but string was assumed. The string we want is
the 1th element of the tuple.

To standardise the device control strings, the parsing was made case-insensitive, so that
ON, OFF and AUTO are accepted.

Updated docstrings to reflect new device control strings (sump and temp-top-up).